+++ FreeAPS/Sources/Services/HealthKit/HealthKitManager.swift
@@ -4,23 +4,16 @@ import HealthKit
 import Swinject
 
 protocol HealthKitManager: GlucoseSource, CarbsSource {
-    /// Check all needed permissions
-    /// Return false if one or more permissions are deny or not choosen
-    var areAllowAllPermissions: Bool { get }
-    /// Check availability to save data of BG type to Health store
-    func checkAvailabilitySaveBG() -> Bool
-    /// Check availability to save data of Carbs type to Health store
-    func checkAvailabilitySaveCarbs() -> Bool
-    /// Requests user to give permissions on using HealthKit
+    /// Requests user to give permissions to using HealthKit
     func requestPermission(completion: ((Bool, Error?) -> Void)?)
     /// Save blood glucose to Health store (dublicate of bg will ignore)
     func saveIfNeeded(bloodGlucose: [BloodGlucose])
     /// Save carbs to Health store (dublicate of bg will ignore)
     func saveIfNeeded(carbs: [CarbsEntry])
-    /// Create observer for data passing beetwen Health Store and FreeAPS
-    func createObserver()
-    /// Enable background delivering objects from Apple Health to FreeAPS
-    func enableBackgroundDelivery()
+    /// Save pumpHistoryEvents (basal and bolus event)
+    func saveIfNeeded(pumpEvents: [PumpHistoryEvent])
+    /// Configure HealthKit manager
+    func configureManager()
     /// Delete glucose with syncID
     func deleteCarbs(syncID: String)
     /// Delete glucose with syncID
@@ -29,12 +22,16 @@ protocol HealthKitManager: GlucoseSource, CarbsSource {
 
 final class BaseHealthKitManager: HealthKitManager, Injectable {
     private enum Config {
-        // unwraped HKObjects
-        static var permissions: Set<HKSampleType> { Set([healthBGObject, healthCarbObject].compactMap { $0 }) }
+        // permissions for write and read
+        static var readPermissions: Set<HKSampleType> { Set([healthBGObject, healthCarbObject].compactMap { $0 }) }
+        static var writePermissions: Set<HKSampleType> {
+            Set([healthBGObject, healthCarbObject, healthInsulinObject].compactMap { $0 })
+        }
 
         // link to object in HealthKit
         static let healthBGObject = HKObjectType.quantityType(forIdentifier: .bloodGlucose)
         static let healthCarbObject = HKObjectType.quantityType(forIdentifier: .dietaryCarbohydrates)
+        static let healthInsulinObject = HKObjectType.quantityType(forIdentifier: .insulinDelivery)
 
         // Meta-data key of FreeASPX data in HealthStore
         static let freeAPSMetaKey = "fromFreeAPSX"
@@ -86,12 +83,6 @@ final class BaseHealthKitManager: HealthKitManager, Injectable {
         HKHealthStore.isHealthDataAvailable()
     }
 
-    var areAllowAllPermissions: Bool {
-        Set(Config.permissions.map { healthKitStore.authorizationStatus(for: $0) })
-            .intersection([.sharingDenied, .notDetermined])
-            .isEmpty
-    }
-
     // NSPredicate, which use during load increment BG from Health store
     private var loadHealthDataPredicate: NSPredicate {
         // loading only daily bg
@@ -116,26 +107,23 @@ final class BaseHealthKitManager: HealthKitManager, Injectable {
         injectServices(resolver)
         guard isAvailableOnCurrentDevice,
               Config.healthBGObject != nil else { return }
+        configureManager()
+        subscribe()
+        debug(.service, "HealthKitManager did create")
+    }
+
+    func configureManager() {
         createObserver()
         enableBackgroundDelivery()
-        debug(.service, "HealthKitManager did create")
-        subscribe()
+        debug(.service, "HealthKitManager did configured")
     }
 
     private func subscribe() {
         broadcaster.register(CarbsObserver.self, observer: self)
     }
 
-    func checkAvailabilitySave(objectTypeToHealthStore: HKObjectType) -> Bool {
+    private func checkAvailabilitySave(objectTypeToHealthStore: HKObjectType) -> Bool {
         healthKitStore.authorizationStatus(for: objectTypeToHealthStore) == .sharingAuthorized
-    }
-
-    func checkAvailabilitySaveBG() -> Bool {
-        Config.healthBGObject.map { checkAvailabilitySave(objectTypeToHealthStore: $0) } ?? false
-    }
-
-    func checkAvailabilitySaveCarbs() -> Bool {
-        Config.healthCarbObject.map { checkAvailabilitySave(objectTypeToHealthStore: $0) } ?? false
     }
 
     func requestPermission(completion: ((Bool, Error?) -> Void)? = nil) {
@@ -143,12 +131,12 @@ final class BaseHealthKitManager: HealthKitManager, Injectable {
             completion?(false, HKError.notAvailableOnCurrentDevice)
             return
         }
-        guard Config.permissions.isNotEmpty else {
+        guard Config.readPermissions.isNotEmpty, Config.writePermissions.isNotEmpty else {
             completion?(false, HKError.dataNotAvailable)
             return
         }
 
-        healthKitStore.requestAuthorization(toShare: Config.permissions, read: Config.permissions) { status, error in
+        healthKitStore.requestAuthorization(toShare: Config.writePermissions, read: Config.readPermissions) { status, error in
             completion?(status, error)
         }
     }
@@ -224,9 +212,89 @@ final class BaseHealthKitManager: HealthKitManager, Injectable {
             .store(in: &lifetime)
     }
 
+    func saveIfNeeded(pumpEvents events: [PumpHistoryEvent]) {
+        guard settingsManager.settings.useAppleHealth,
+              let sampleType = Config.healthInsulinObject,
+              checkAvailabilitySave(objectTypeToHealthStore: sampleType),
+              events.isNotEmpty
+        else { return }
+
+        loadSamplesFromHealth(sampleType: sampleType, withIDs: events.map(\.id))
+            .receive(on: processQueue)
+            .compactMap { samples -> ([InsulinBolus], [InsulinBasal]) in
+                let sampleIDs = samples.compactMap(\.syncIdentifier)
+                let bolus = events
+                    .filter { $0.type == .bolus && !sampleIDs.contains($0.id) }
+                    .compactMap { event -> InsulinBolus? in
+                        guard let amount = event.amount else { return nil }
+                        return InsulinBolus(id: event.id, amount: amount, date: event.timestamp)
+                    }
+                let basalEvents = events
+                    .filter { $0.type == .tempBasal && !sampleIDs.contains($0.id) }
+                let basal = basalEvents.enumerated()
+                    .compactMap { item -> InsulinBasal? in
+                        let nextElementEventIndex = item.offset + 1
+                        guard basalEvents.count > nextElementEventIndex else { return nil }
+                        let nextBasalEvent = basalEvents[nextElementEventIndex]
+                        let secondsOfCurrentBasal = nextBasalEvent.timestamp.timeIntervalSince(item.element.timestamp)
+                        let amount = Decimal(secondsOfCurrentBasal / 3600) * (item.element.rate ?? 0)
+                        let id = String(item.element.id.dropFirst())
+                        guard amount > 0,
+                              id != ""
+                        else { return nil }
+                        return InsulinBasal(
+                            id: id,
+                            amount: amount,
+                            startDelivery: item.element.timestamp,
+                            endDelivery: nextBasalEvent.timestamp
+                        )
+                    }
+                return (bolus, basal)
+            }
+            .sink(receiveValue: { bolus, basal in
+                // save bolus
+                let bolusSamples = bolus
+                    .map {
+                        HKQuantitySample(
+                            type: sampleType,
+                            quantity: HKQuantity(unit: .internationalUnit(), doubleValue: Double($0.amount)),
+                            start: $0.date,
+                            end: $0.date,
+                            metadata: [
+                                HKMetadataKeyInsulinDeliveryReason: NSNumber(2),
+                                HKMetadataKeyExternalUUID: $0.id,
+                                HKMetadataKeySyncIdentifier: $0.id,
+                                HKMetadataKeySyncVersion: 1,
+                                Config.freeAPSMetaKey: true
+                            ]
+                        )
+                    }
+
+                let basalSamples = basal
+                    .map {
+                        HKQuantitySample(
+                            type: sampleType,
+                            quantity: HKQuantity(unit: .internationalUnit(), doubleValue: Double($0.amount)),
+                            start: $0.startDelivery,
+                            end: $0.endDelivery,
+                            metadata: [
+                                HKMetadataKeyInsulinDeliveryReason: NSNumber(1),
+                                HKMetadataKeyExternalUUID: $0.id,
+                                HKMetadataKeySyncIdentifier: $0.id,
+                                HKMetadataKeySyncVersion: 1,
+                                Config.freeAPSMetaKey: true
+                            ]
+                        )
+                    }
+
+                self.healthKitStore.save(bolusSamples + basalSamples) { _, _ in }
+            })
+            .store(in: &lifetime)
+    }
+
     // MARK: - Observers & Background data delivery
 
-    func createObserver() {
+    private func createObserver() {
         guard settingsManager.settings.useAppleHealth else { return }
 
         createBGObserver()
@@ -277,7 +345,7 @@ final class BaseHealthKitManager: HealthKitManager, Injectable {
         debug(.service, "Create Observer for Carbs")
     }
 
-    func enableBackgroundDelivery() {
+    private func enableBackgroundDelivery() {
         guard settingsManager.settings.useAppleHealth else {
             healthKitStore.disableAllBackgroundDelivery { _, _ in }
             return }
@@ -591,4 +659,19 @@ enum HKError: Error {
     case notAvailableOnCurrentDevice
     // Some data can be not available on current iOS-device
     case dataNotAvailable
+}
+
+// MARK: - Subtypes
+
+private struct InsulinBolus {
+    var id: String
+    var amount: Decimal
+    var date: Date
+}
+
+private struct InsulinBasal {
+    var id: String
+    var amount: Decimal
+    var startDelivery: Date
+    var endDelivery: Date
 }